+++ packages/cli/src/commands/split/utils/extract-filename-from-path.ts
@@ -1,0 +1,5 @@
+import * as path from 'node:path';
+
+export function extractFileNameFromPath(filename: string) {
+  return path.basename(filename, path.extname(filename));
+}