+++ packages/cli/src/commands/split/utils/get-file-name-path.ts
@@ -1,0 +1,5 @@
+import * as path from 'node:path';
+
+export function getFileNamePath(componentDirPath: string, componentName: string, ext: string) {
+  return path.join(componentDirPath, componentName) + `.${ext}`;
+}