+++ packages/cli/src/commands/stats/index.ts
@@ -7,74 +7,16 @@ import {
   normalizeVisitors,
   walkDocument,
   bundle,
-  logger,
-  type OASStatsAccumulator,
-  type AsyncAPIStatsAccumulator,
   type WalkContext,
   type OutputFormat,
 } from '@redocly/openapi-core';
-import * as colors from 'colorette';
 import { performance } from 'perf_hooks';
 
 import type { VerifyConfigOptions } from '../../types.js';
-import { getFallbackApisOrExit, printExecutionTime } from '../../utils/miscellaneous.js';
+import { getFallbackApisOrExit } from '../../utils/miscellaneous.js';
 import type { CommandArgs } from '../../wrapper.js';
+import { printStats } from './print-stats/index.js';
 import { resolveStatsVisitorAndAccumulator } from './visitor-and-accumulator-resolver.js';
-
-function printStatsStylish(statsAccumulator: OASStatsAccumulator | AsyncAPIStatsAccumulator) {
-  for (const node in statsAccumulator) {
-    const stat = statsAccumulator[node as keyof typeof statsAccumulator];
-    const { metric, total, color } = stat;
-    const colorFn = colors[color as keyof typeof colors] as (text: string) => string;
-    logger.output(colorFn(`${metric}: ${total} \n`));
-  }
-}
-
-function printStatsJson(statsAccumulator: OASStatsAccumulator | AsyncAPIStatsAccumulator) {
-  const json: any = {};
-  for (const key of Object.keys(statsAccumulator)) {
-    const stat = statsAccumulator[key as keyof typeof statsAccumulator];
-    json[key] = {
-      metric: stat.metric,
-      total: stat.total,
-    };
-  }
-
-  logger.output(JSON.stringify(json, null, 2));
-}
-
-function printStatsMarkdown(statsAccumulator: OASStatsAccumulator | AsyncAPIStatsAccumulator) {
-  let output = '| Feature  | Count  |\n| --- | --- |\n';
-  for (const key of Object.keys(statsAccumulator)) {
-    const stat = statsAccumulator[key as keyof typeof statsAccumulator];
-    output += '| ' + stat.metric + ' | ' + stat.total + ' |\n';
-  }
-
-  logger.output(output);
-}
-
-function printStats(
-  statsAccumulator: OASStatsAccumulator | AsyncAPIStatsAccumulator,
-  api: string,
-  startedAt: number,
-  format: string
-) {
-  logger.info(`Document: ${colors.magenta(api)} stats:\n\n`);
-
-  switch (format) {
-    case 'stylish':
-      printStatsStylish(statsAccumulator);
-      break;
-    case 'json':
-      printStatsJson(statsAccumulator);
-      break;
-    case 'markdown':
-      printStatsMarkdown(statsAccumulator);
-      break;
-  }
-
-  printExecutionTime('stats', startedAt, api);
-}
 
 export type StatsArgv = {
   api?: string;