+++ packages/respect-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @redocly/respect-core
 
+## 2.26.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.26.0.
+
 ## 2.25.4
 
 ### Patch Changes